// src/api/generated/model/document.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Fiber OpenAPI
 * API documentation generated by fiber-oapi
 * OpenAPI spec version: 1.0.0
 */
import type { Block } from './block';
import type { Document as __Document } from './document';
import type { DocumentConfig } from './documentConfig';
import type { DocumentMetadata } from './documentMetadata';
import type { Space } from './space';
import type { Time } from './time';

export interface Document {
  config?: DocumentConfig;
  content?: Block[];
  created_at?: Time;
  id?: string;
  metadata?: DocumentMetadata;
  name?: string;
  parent?: __Document;
  parent_id?: string;
  slug?: string;
  space?: Space;
  space_id?: string;
  updated_at?: Time;
}